requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord==2.2.3
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
